conv MOS to rreil & tests

+++ lib/src/mos/load.rs
@@ -78,7 +78,7 @@ pub fn load(p: &Path) -> Result<Project,Error> {
     let mut prog = Program::new("prog0", Target::Mos6502);
     let mut proj = Project::new(name.clone(),reg);
 
-    prog.call_graph.add_vertex(CallTarget::Todo(Rvalue::new_u16(addr),Some(name),Uuid::new_v4()));
+    prog.call_graph.add_vertex(CallTarget::Todo(Rvalue::new_u16(addr as u16),Some(name),Uuid::new_v4()));
     proj.comments.insert(("base".to_string(),addr),"main".to_string());
 
     for &(name,ref off,cmnt) in Target::Mos6502.interrupt_vec().iter() {
